pip3_requirements.txt
actionlib==1.11.16
angles==1.9.12
apturl==0.5.2
base-local-planner==1.14.9
beautifulsoup4==4.4.1
blinker==1.3
bondpy==1.8.3
Brlapi==0.6.4
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
capabilities==0.2.0
catkin==0.7.29
catkin-pkg-modules==0.4.24
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
create-dashboard==2.3.1
create-driver==2.3.1
create-node==2.3.1
cryptography==1.2.3
cv-bridge==1.12.8
cycler==0.11.0
defer==1.0.6
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distro-info===0.14ubuntu0.2
docutils==0.12
dynamic-reconfigure==1.5.50
feedparser==5.1.3
gazebo-plugins==2.5.21
gazebo-ros==2.5.21
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.14
guacamole==0.9.2
html5lib==0.999
httplib2==0.9.1
idna==2.0
image-geometry==1.12.8
interactive-markers==1.11.5
Jinja2==2.8
joint-state-publisher==1.12.15
kiwisolver==1.4.2
kobuki-dashboard==0.5.7
language-selector==0.1
laser-geometry==1.6.7
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
matplotlib==3.4.0
message-filters==1.12.17
numpy==1.19.0
oauthlib==1.0.3
onboard==1.2.0
openni2-launch==1.5.1
padme==1.1.1
pexpect==4.0.1
Pillow==8.3.1
plainbox==0.25
pluginlib==1.11.3
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==3.0.8
python-apt==1.1.0b1+ubuntu0.16.4.12
python-dateutil==2.8.2
python-debian==0.1.27
python-qt-binding==0.3.8
python-systemd==231
pyxdg==0.25
PyYAML==3.11
qt-dotgraph==0.3.18
qt-gui==0.3.18
qt-gui-cpp==0.3.18
qt-gui-py-common==0.3.18
reportlab==3.3.0
requests==2.9.1
resource-retriever==1.12.6
rocon-app-manager==0.9.1
rocon-app-utilities==0.9.1
rocon-apps==0.9.1
rocon-console==0.3.2
rocon-ebnf==0.3.2
rocon-gateway==0.8.1
rocon-gateway-utils==0.8.1
rocon-hub==0.8.1
rocon-hub-client==0.8.1
rocon-interactions==0.3.2
rocon-master-info==0.3.2
rocon-python-comms==0.3.2
rocon-python-redis==0.3.2
rocon-python-utils==0.3.2
rocon-python-wifi==0.3.2
rocon-semantic-version==0.3.2
rocon-uri==0.3.2
roman==2.0.0
rosbag==1.12.17
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.12.17
roslaunch==1.12.17
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.17
rosmake==1.14.6
rosmaster==1.12.17
rosmsg==1.12.17
rosnode==1.12.17
rosparam==1.12.17
rospkg-modules==1.3.0
rospy==1.12.17
rospy-message-converter==0.5.6
rosserial-python==0.7.7
rosservice==1.12.17
rostest==1.12.17
rostopic==1.12.17
rosunit==1.14.6
roswtf==1.12.17
rqt-action==0.4.9
rqt-bag==0.5.0
rqt-bag-plugins==0.5.0
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.16
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.12.17
screen-resolution-extra==0.0.0
sensor-msgs==1.12.7
sessioninstaller==0.0.0
six==1.16.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
ssh-import-id==5.5
system-service==0.3
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.17
torch==1.8.1+cu101
torchvision==0.9.1+cu101
turtlebot3-automatic-parking==1.1.0
turtlebot3-automatic-parking-vision==1.1.0
turtlebot3-autorace-camera==1.2.0
turtlebot3-autorace-control==1.2.0
turtlebot3-autorace-core==1.2.0
turtlebot3-autorace-detect==1.2.0
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
typing-extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unique-id==1.0.5
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
UNKNOWN==9.0.1
urllib3==1.13.1
usb-creator==0.3.0
world-canvas-server==0.2.0
world-canvas-utils==0.2.0
xacro==1.11.3
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
